+++ assets/js/wpdm-customization-images-admin.js
@@ -109,3 +109,4 @@ jQuery(document).ready(function($) {
 
 
 
+