+++ includes/class-wpdm-customization-images-admin.php
@@ -304,3 +304,4 @@ class WPDM_Customization_Images_Admin {
 
 
 
+